refactor(tests): split RFC/Epic tests into separate file
- Move RFC/Epic label tests to test_task_status_rfc_epic.py
- Reduce test_task_status_dashboard.py from 542 to 369 lines
- Fix CI block threshold violation (>400 LOC)
- All 13 tests pass, no functionality changes

Resolves CI failure from PR #1499.
